backend/requirements.txt
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-import-export==4.3.9
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.16.0
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
